python-service/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pdfplumber==0.11.4
python-docx==1.1.2
spacy==3.8.4
